+++ src/main/java/com/buuz135/industrial/api/fluid/IFluidDrinkHandler.java
@@ -7,11 +7,10 @@ import net.minecraftforge.fluids.FluidStack;
 
 public interface IFluidDrinkHandler {
     /**
-     *
-     * @param world World of where drinking
-     * @param pos BlockPos of where drinking from
-     * @param stack FluidStack drank
-     * @param player EntityPlayer drinking
+     * @param world              World of where drinking
+     * @param pos                BlockPos of where drinking from
+     * @param stack              FluidStack drank
+     * @param player             EntityPlayer drinking
      * @param fromFluidContainer true if drinking from a tank
      */
     void onDrink(World world, BlockPos pos, FluidStack stack, EntityPlayer player, boolean fromFluidContainer);