+++ src/main/java/com/buuz135/industrial/tile/CustomElectricMachine.java
@@ -47,5 +47,8 @@ public abstract class CustomElectricMachine extends ElectricMachine {
         return 0;
     }
 
-
+    @Override
+    protected int getMinimumWorkTicks() {
+        return 4;
+    }
 }