+++ src/main/java/com/buuz135/industrial/utils/StrawUtils.java
@@ -7,16 +7,17 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class StrawUtils {
-    public static final Map<String,IFluidDrinkHandler> DRINK_HANDLER_MAP = new HashMap<>();
+    public static final Map<String, IFluidDrinkHandler> DRINK_HANDLER_MAP = new HashMap<>();
+
+    private StrawUtils() {
+    }
 
     public static void register(Fluid fluid, IFluidDrinkHandler handler) {
         register(fluid.getName(), handler);
     }
+
     public static void register(String fluid, IFluidDrinkHandler handler) {
         DRINK_HANDLER_MAP.put(fluid, handler);
-    }
-
-    private StrawUtils() {
     }
 
     public static Map<String, IFluidDrinkHandler> getDrinkHandlers() {